告警修复

+++ interfaces/kits/cj/src/media_asset_manager_ffi.cpp
@@ -49,7 +49,7 @@ const int32_t HIGH_QUALITY_IMAGE = 0;
 const int32_t UUID_STR_LENGTH = 37;
 const int32_t MAX_URI_SIZE = 384; // 256 for display name and 128 for relative path
 const int32_t REQUEST_ID_MAX_LEN = 64;
-const ssize_t MAX_IMAGE_BUFFER_SIZE = 100 * 1024 * 1024; // 100MB limit for image buffer
+const int64_t MAX_IMAGE_BUFFER_SIZE = 100 * 1024 * 1024; // 100MB limit for image buffer
 
 static mutex multiStagesCaptureLock;
 static mutex registerTaskLock;
